made it work
